+++ lib/prawn-rails/config.rb
@@ -5,13 +5,8 @@ module PrawnRails
 
   @config = OpenStruct.new(page_layout: :portrait, page_size: "A4", skip_page_creation: false)
 
-  def config
-    begin
-      block_given? ? yield(@config) : @config
-    rescue => e
-      puts e
-      puts e.backtrace
-    end
+  def config(&block)
+    block_given? ? yield(@config) : @config
   end
 
 end