+++ test/dummy_app/config/application.rb
@@ -8,6 +8,10 @@ require "prawn-rails"
 
 module Dummy
   class Application < Rails::Application
+    if Rails::VERSION::STRING.to_f >= 5.1
+      config.load_defaults Rails::VERSION::STRING.to_f
+    end
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.