+++ test/dummy_app/config/initializers/prawn-rails.rb
@@ -1,5 +1,0 @@
-PrawnRails.config do |config|
-  config.page_layout = :portrait
-  config.page_size   = "A4"
-  config.skip_page_creation = false
-end